implemented backup saving to json

+++ PDBbind_data/similarity/pairwise_similarity_matrix/pairwise_similarity_tm_rmsd.py
@@ -1,3 +1,4 @@
+import random
 import argparse
 from scipy.spatial import KDTree
 import json
@@ -22,7 +23,7 @@ from mpl_toolkits.mplot3d import Axes3D
 
 # Set up logging with rotation
 log_formatter = logging.Formatter('%(asctime)s %(levelname)s %(message)s')
-log_file = 'pairwise_similarity_tm_rmsd_error_8.log'
+log_file = f'pairwise_similarity_tm_rmsd_error.log'
 
 # Set up a rotating file handler
 rotating_handler = RotatingFileHandler(log_file, mode='a', maxBytes=5*1024*1024, backupCount=3, encoding=None, delay=0)
@@ -132,24 +133,20 @@ def run_tmalign(tm_align_path, pdb1, pdb2, matrix_path=None):
 
 
 
-
 def point_cloud_similarity_score(pc1, pc2):
-    # Ensure both point clouds are numpy arrays
     pc1 = np.array(pc1)
     pc2 = np.array(pc2)
     
-    # Build a KD-tree for the second point cloud
-    tree = KDTree(pc2)
+    # Build KD-trees for both point clouds
+    tree1 = KDTree(pc1)
+    tree2 = KDTree(pc2)
     
-    # Find the nearest neighbors for each point in pc1
-    distances, _ = tree.query(pc1)
-
-    # Compute the root mean squared deviation
-    rmsd = np.sqrt(np.mean(distances**2))
+    # Compute distances
+    distances1, _ = tree2.query(pc1)
+    distances2, _ = tree1.query(pc2)
     
-    return rmsd
-
-
+    # Compute the symmetric root mean squared deviation using max
+    return max(np.sqrt(np.mean(distances1**2)), np.sqrt(np.mean(distances2**2)))
 
 
 
@@ -162,13 +159,11 @@ def process_pair(id1, id2, mol1, mol2, folder_path, tm_align_path):
     try:
 
         # Align proteins with TM-align
-
         pdb1 = os.path.join(folder_path, f"{id1}.pdb")
         pdb2 = os.path.join(folder_path, f"{id2}.pdb")
         
         tm_score, seq_id = run_tmalign(tm_align_path, pdb1, pdb2, matrix_path)
         
-
         # Compute ligand positioning similarity
         rot_matrix, t_vector = parse_rotation_matrix_and_translation_vector(matrix_path)
 
@@ -185,7 +180,8 @@ def process_pair(id1, id2, mol1, mol2, folder_path, tm_align_path):
         if os.path.exists(matrix_path): os.remove(matrix_path)
 
     except Exception as e:
-        logger.error(f"Error processing pair {id1}, {id2}: {str(e)}")
+        if os.path.exists(matrix_path): os.remove(matrix_path)
+        logger.error(f"--- Error processing pair {id1}, {id2}: {str(e)}")
 
     return [tm_score, ligand_rmsd]
 
@@ -195,33 +191,30 @@ def process_pair(id1, id2, mol1, mol2, folder_path, tm_align_path):
 
 
 
-def main(folder_path, tm_align_path, start_complex=0, end_complex=19443):
+def main(folder_path, tm_align_path, save_as_json=True):
 
     try:
-        run = 8
-
-        # List of the names of the complexes
         complexes = sorted([compl[0:4] for compl in os.listdir(folder_path) 
                             if compl[0].isdigit() and compl.endswith('.pdb')])
-
         num_complexes = len(complexes)
         print("Number of complexes: {}".format(num_complexes), flush=True)
 
         # Save list of complexes to json file
-        with open('pairwise_similarity_complexes.json', 'w') as f:
-            json.dump(complexes, f)
-        print("List of complexes saved to pairwise_similarity_complexes.json", flush=True)
+        if not os.path.exists('pairwise_similarity_complexes.json'):
+            with open('pairwise_similarity_complexes.json', 'w') as f:
+                json.dump(complexes, f)
+            print("List of complexes saved to pairwise_similarity_complexes.json", flush=True)
 
 
         # Initialize the HDF5 file and dataset to save the similarities
-        with h5py.File(f'pairwise_similarity_tm_scores_{run}.hdf5', 'a') as f:
+        with h5py.File(f'pairwise_similarity_tm_scores.hdf5', 'a') as f:
             if 'similarities' not in f:
                 dset = f.create_dataset("similarities", (num_complexes, num_complexes), dtype='float32', compression="gzip")
             else:
                 dset = f['similarities']
 
 
-        with h5py.File(f'pairwise_similarity_rmsd_ligand_{run}.hdf5', 'a') as f:
+        with h5py.File(f'pairwise_similarity_rmsd_ligand.hdf5', 'a') as f:
             if 'similarities' not in f:
                 dset = f.create_dataset("similarities", (num_complexes, num_complexes), dtype='float32', compression="gzip")
             else:
@@ -229,50 +222,109 @@ def main(folder_path, tm_align_path, start_complex=0, end_complex=19443):
 
 
         # Parse the SDF files and store the molecules in a dictionary
-        print("Parsing all SDF files...", flush=True)
+        print("Parsing all SDF files...")
         parsed_molecules = parse_sdf_files(folder_path, complexes)
-        print("Parsed all ligands!", flush=True)
+        print("Parsed all ligands!")
+        print()
 
 
         # Loop through each complex and compare it with all other complexes in parallel
         tic = time()
         for i in range(num_complexes):
 
-            if not start_complex <= i < end_complex: continue
+            print(f"Processing {complexes[i]} ({i})...", flush=True)
 
-            to_compare = [j for j in range(i+1, num_complexes)]
-            if len(to_compare) == 0: continue
+            # Check if the complex's similarities have already been precomputed
+            precomputed_rmsds = os.path.join(folder_path, f"{complexes[i]}_similarities_rmsd.json")
+            precomputed_tm_scores = os.path.join(folder_path, f"{complexes[i]}_similarities_tm_scores.json")
+            
+            if os.path.exists(precomputed_rmsds) and os.path.exists(precomputed_tm_scores):
+                
+                try:
+                    # Append RMSD data to the h5py file
+                    with open(precomputed_rmsds, 'r') as f:
+                        rmsd_data = json.load(f)
+                        rmsds = [rmsd_data[complex] for complex in complexes] # Convert values to list in the right order
 
-            # RUN ALL THE COMPARISONS IN PARALLEL, accumulate the results
-            results = Parallel(n_jobs=-1)(delayed(process_pair)(
-                    complexes[i], complexes[j], parsed_molecules[complexes[i]], parsed_molecules[complexes[j]],
-                    folder_path, tm_align_path) for j in to_compare)
+                    with h5py.File(f'pairwise_similarity_rmsd_ligand.hdf5', 'a') as f:
+                            dset = f['similarities']
+                            dset[i, :] = rmsds
+                            dset[:, i] = rmsds
+
+                    # Append TM scores data to the h5py file
+                    with open(precomputed_tm_scores, 'r') as f:
+                        tm_data = json.load(f)
+                        tm_scores = [tm_data[complex] for complex in complexes] # Convert values to list in the right order
+                    with h5py.File(f'pairwise_similarity_tm_scores.hdf5', 'a') as f:
+                            dset = f['similarities']
+                            dset[i, :] = tm_scores
+                            dset[:, i] = tm_scores
+                    print(f"--- Precomputed data found for {complexes[i]}, skipping...", flush=True)
+                
+                except Exception as e:
+                    logger.error(f"Error in retrieving precomputed data: {str(e)}")
 
 
-            # APPEND THE TM SCORES TO THE HDF5 FILE
-            with h5py.File(f'pairwise_similarity_tm_scores_{run}.hdf5', 'a') as f:
-                dset = f['similarities']
-                for j, metrics in zip(to_compare, results):
-                    dset[i, j] = metrics[0]
-                    dset[j, i] = metrics[0]
+            # Compare complex to all other complexes in parallel
+            else:
+                to_compare = [j for j in range(i+1, num_complexes)]
+                if len(to_compare) == 0: continue
 
-            # APPEND THE LIGAND RMSD TO THE HDF5 FILE
-            with h5py.File(f'pairwise_similarity_rmsd_ligand_{run}.hdf5', 'a') as f:
+                print(f"--- Comparing {complexes[i]} ({i}) to indexes {to_compare[0]}-{to_compare[-1]}", flush=True)
+
+                tac = time()
+                # RUN ALL THE COMPARISONS IN PARALLEL, accumulate the results
+                results = Parallel(n_jobs=-1)(delayed(process_pair)(
+                        complexes[i], complexes[j], parsed_molecules[complexes[i]], parsed_molecules[complexes[j]],
+                        folder_path, tm_align_path) for j in to_compare)
+
+
+                # APPEND THE TM SCORES TO THE HDF5 FILE
+                with h5py.File(f'pairwise_similarity_tm_scores.hdf5', 'a') as f:
                     dset = f['similarities']
                     for j, metrics in zip(to_compare, results):
-                        dset[i, j] = metrics[1]
-                        dset[j, i] = metrics[1]
-            
-            print(f"Time: {time() - tic:.2f} - Compared {complexes[i]} ({i}) to indexes {to_compare[0]}-{to_compare[-1]}", flush=True)
+                        dset[i, j] = metrics[0]
+                        dset[j, i] = metrics[0]
+
+                    if save_as_json:
+
+                        # Extract line from the dataset and save it to a JSON file
+                        line_data = np.array(dset[i, :], dtype=np.float32)
+                        line_data = line_data.tolist()
+                        sim_data = {complexes[i]: line_data[i] for i in range(len(complexes))}
+                    
+                        # Save the dictionary to a JSON file with indentation
+                        with open(precomputed_tm_scores, 'w') as f:
+                            json.dump(sim_data, f, indent=4)
+
+
+                # APPEND THE LIGAND RMSD TO THE HDF5 FILE
+                with h5py.File(f'pairwise_similarity_rmsd_ligand.hdf5', 'a') as f:
+                        dset = f['similarities']
+                        for j, metrics in zip(to_compare, results):
+                            dset[i, j] = metrics[1]
+                            dset[j, i] = metrics[1]
+
+                        if save_as_json:
+
+                            # Extract line from the dataset and save it to a JSON file
+                            line_data = np.array(dset[i, :], dtype=np.float32)
+                            line_data = line_data.tolist()
+                            sim_data = {complexes[i]: line_data[i] for i in range(len(complexes))}
+                        
+                            # Save the dictionary to a JSON file with indentation
+                            with open(precomputed_rmsds, 'w') as f:
+                                json.dump(sim_data, f, indent=4)
+                
+                toc = time()
+                print(f"--- Done: Time needed: {toc - tac:.2f} seconds. Total Time: {toc - tic}", flush=True)
 
 
     except Exception as e:
         logger.error(f"Error in main function: {str(e)}")
         sys.exit(1)
 
-    toc = time()
-    print("Elapsed time: {:.2f} seconds".format(toc - tic))
-
+    print(f"Elapsed time: {time() - tic:.2f} seconds", flush=True)
 
 
 
@@ -281,9 +333,7 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser(description="Compute and store pairwise metrics for 3D complexes.")
     parser.add_argument('folder_path', type=str, help='Path to the folder containing the 3D complexes')
     parser.add_argument('tm_align_path', type=str, help='Path to the TM-align executable')
-    parser.add_argument('start_complex', type=int, default=0, help='Index of the first complex to process')
-    parser.add_argument('end_complex', type=int, default=19443, help='Index of the last complex to process')
     args = parser.parse_args()
 
-    main(args.folder_path, args.tm_align_path, args.start_complex, args.end_complex)
+    main(args.folder_path, args.tm_align_path)
